Remove static storage duration for dad_count
create_bound_listener is called by set_interfaces() which can configure
the set of interfaces to listen on at runtime. Since dad_count was
static, this means that the DAD delay only took effect for the first
call to set_interfaces(). All subsequent calls would not retry bind()
but just abort() on failure.

This change removes the static keyword from dad_count to change it to
use automatic storage duration instead, which properly resets the value
to 0 every time the function is run.

There is a slight side effect that if multiple interfaces are
configured, create_bound_listener can now sleep for every interface,
instead of awaiting the dad delay a single time. A better approach would
be passing around the current dad_count, but given this code is
effectively unmaintained and will be replaced in the not-so-distant
future, this change is much more straight-forward.

Bug: 451883641
Bug: 444145492
Test: TH
Flag: NONE yes
Change-Id: I0bc568b0001ea6614fe765ecceebd740432af3e1